Add an API to fetch repositories and list their branches.
This API is a replacement for the one currently provided by
outpack_server.

A single instance of orderly.runner API is able to operate over an
arbitrary number of Git repositories. The first time a URL is fetched, a
`git clone` is performed. Subsequent fetches are incremental using `git
fetch`.

Repositories are each stored in their own directory, named using the
hash of the URL. Using the hash makes it easy to support arbitrary
upstreams, without having to worry about encoding the URL characters as
file names.

The existing endpoints still operate on the "root" repository, which is
assumed to exist before the API is launched. Eventually all those APIs
will be migrated to have a `url` argument.