Decouple the runner from the outpack_server store.
In the previous design, outpack_server, the runner API and the workers
all shared a single outpack directory, and a Git repository (mostly, the
workers used a clone of the shared repository for the actual execution).

This creates a very tight and brittle coupling between all the
components. It makes it impossible to deploy the different components on
separate machines. It requires careful reasoning about data races and
conflicts between the different bits. It prevents us from sharing worker
processes across multiple Packit instances, and it prevents us from
using multiple Git repositories within a single instance.

The new design completely splits up the storage.
- The API server and each worker have their own local Git clones of the
  repositories, that are directly pulled from the upstream (eg. GitHub).
- The API servers and workers store bare Git clones of the repositories,
  without any worktree. When running a report, workers create a new
  worktree in a temporary directory, run the report and delete the
  worktree. This ensures a completely clean slate every time.
- The workers use their own outpack store, that is not shared with any
  other process.
- The workers can pull and push packets using any protocol supported by
  orderly2. In practice, we will be using HTTP to interact with the
  outpack_server used by Packit.

Currently, the workers create a new outpack store for each run, meaning
they do not cache any of the packet dependencies and need to download
them from the outpack_server from scratch every time. Given that, at
least for now, workers and outpack_server will be operating on the same
or nearby machines, this seems like a reasonable overhead.

Ideally we would keep a per-worker cache, however we need to be careful
not to mix packets between different instances. One possible approach
may be to re-use the file store, but start from an empty metadata store
everytime. This way large unnecessary file downloads are avoided, while
preserving some degree of isolation between runs and instances.