Use entire api object in API tests.
The API tests used to setup individual endpoints to run the tests. That
is a bit verbose and duplicates a bunch of code from the `api.R` file.

This will be made much worse when we introduce `url` parameters to all
the endpoints, since all the tests will need to make a API call to fetch
the repository first, which will lead to more boilerplate.

Using the api object solves this issue.