Fix the reported modification time of reports.
The `/report/list` endpoint uses the `git_ls` function to enumerate
reports, and supposedly includes the modification time is the response.

Unfortunately the modification time returned by `git_ls` is pretty
useless, and actually refers to the modification times in the local
worktree, which has nothing to do with commit times.

The proper way to do this is to identify the last commit to have touched
the file, and then get the time of that commit. Unfortunately gert does
not have an easy way to list commits on just one file (and in fact
neither does libgit2), so we fallback to using the `git rev-list`
command.

Additionally I have re-written the code that was used to compare reports
against the default branch to use Git plumbing commands instead of the
high-level porcelain ones.