Include R errors in the log files.
Currently, when an R error is thrown by a report, it is caught by the
rrq worker and is stored in Redis, but it is not exposed over the runner
API anywhere.

Rather than introduce yet another field in API, we print the error from
the worker process, which will make it visible at the end of the tasks
log file.