Use the metadata-action to set Docker labels.
For packit-infra's automatic nightly update action, I want to be able to
print a list of commits associated with the update. This isn't currently
possible as there's no way to know from a Docker image the Git revision
that was used to build it.

Docker labels allow us to encode extra bits of metadata in the image,
such as this. Rather than re-implement all the metadata extraction and
formatting, we now use the docker/metadata-action action which takes
care of it all. This also replaces the old bash script that was
calculating the tags to add to the image. The action needs a bit of
customization to match the existing tags.

This also add `latest` as a tag for builds from the default branch, in
addition to labeling them with the branch name.